# atm-service/src/main/resources/swagger/bank/bank-swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Api Documentation",
    "version": "1.0",
    "title": "Api Documentation",
    "termsOfService": "urn:tos",
    "contact": {},
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "host": "localhost:8090",
  "basePath": "/",
  "tags": [
    {
      "name": "bank-account-controller",
      "description": "Bank Account Controller"
    }
  ],
  "paths": {
    "/api/account/card": {
      "get": {
        "tags": [
          "bank-account-controller"
        ],
        "summary": "Get card details for given card number",
        "operationId": "getCardUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "cardNumber",
            "in": "query",
            "description": "cardNumber",
            "required": true,
            "type": "string",
            "allowEmptyValue": false
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/CardDto"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "deprecated": false
      }
    },
    "/api/account/card/balance": {
      "get": {
        "tags": [
          "bank-account-controller"
        ],
        "summary": "Get current balance for given card number",
        "operationId": "getCardBalanceUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "cardNumber",
            "in": "query",
            "description": "cardNumber",
            "required": true,
            "type": "string",
            "allowEmptyValue": false,
            "x-example": "12345678"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "number"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "deprecated": false
      }
    },
    "/api/account/card/deposit": {
      "post": {
        "tags": [
          "bank-account-controller"
        ],
        "summary": "Deposit money for given card number",
        "operationId": "depositCardBalanceUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "amount",
            "in": "query",
            "description": "amount",
            "required": true,
            "type": "number",
            "allowEmptyValue": false,
            "x-example": 22.12
          },
          {
            "name": "cardNumber",
            "in": "query",
            "description": "cardNumber",
            "required": true,
            "type": "string",
            "allowEmptyValue": false,
            "x-example": "12345678"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "number"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "deprecated": false
      }
    },
    "/api/account/card/validate": {
      "post": {
        "tags": [
          "bank-account-controller"
        ],
        "summary": "Validate card secret for given card number",
        "operationId": "validateCardSecretUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "cardNumber",
            "in": "query",
            "description": "cardNumber",
            "required": true,
            "type": "string",
            "allowEmptyValue": false,
            "x-example": "12345678"
          },
          {
            "name": "preferredAuth",
            "in": "query",
            "description": "preferredAuth",
            "required": true,
            "type": "string",
            "allowEmptyValue": false,
            "enum": [
              "PIN",
              "FINGERPRINT"
            ]
          },
          {
            "name": "secret",
            "in": "query",
            "description": "secret",
            "required": true,
            "type": "string",
            "allowEmptyValue": false,
            "x-example": "1234"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "deprecated": false
      }
    },
    "/api/account/card/withdraw": {
      "post": {
        "tags": [
          "bank-account-controller"
        ],
        "summary": "Withdraw money for given card number",
        "operationId": "withdrawCardBalanceUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "amount",
            "in": "query",
            "description": "amount",
            "required": true,
            "type": "number",
            "allowEmptyValue": false,
            "x-example": 11.2
          },
          {
            "name": "cardNumber",
            "in": "query",
            "description": "cardNumber",
            "required": true,
            "type": "string",
            "allowEmptyValue": false,
            "x-example": "12345678"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "number"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "deprecated": false
      }
    }
  },
  "securityDefinitions": {
    "Api key": {
      "type": "apiKey",
      "name": "X-API-KEY",
      "in": "header"
    }
  },
  "definitions": {
    "CardDto": {
      "type": "object",
      "properties": {
        "balance": {
          "type": "number"
        },
        "cardNumber": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "preferredAuth": {
          "type": "string",
          "enum": [
            "PIN",
            "FINGERPRINT"
          ]
        }
      },
      "title": "CardDto"
    }
  }
}